+++ Cargo.toml
@@ -21,8 +21,8 @@ same-file = "1.0.6"
 remove_dir_all = "0.7"
 #jemallocator = "0.3.2"
 regex = "1.5"
-sl-sh-proc-macros = { git = "https://github.com/sl-sh-dev/sl-sh-proc-macros.git" }
-#sl-sh-proc-macros = { path = "../sl-sh-proc-macros" }
+#sl-sh-proc-macros = { git = "https://github.com/sl-sh-dev/sl-sh-proc-macros.git" }
+sl-sh-proc-macros = { path = "../sl-sh-proc-macros" }
 static_assertions = "1"
 
 [build-dependencies]
