+++ src/builtins_util.rs
@@ -3,8 +3,6 @@ use crate::eval::*;
 use crate::types::*;
 use std::borrow::Cow;
 
-use crate::LispResult;
-use std::cmp::Ordering;
 use std::convert::{TryFrom, TryInto};
 use std::env;
 
@@ -553,14 +551,9 @@ macro_rules! try_inner_string {
     }};
 }
 
-#[derive(Debug, Clone)]
-pub enum ArgType {
-    Exp(Expression),
-    Opt(Option<Expression>),
-    VarArgs(Vec<Expression>),
-}
-
-//TODO #3 can we also support slices?
+/// Used by sl_sh_fn macro to embed information at runtime about the parameters of
+/// the rust native function, specifically whether it is a normal Type, or some
+/// supported wrapped type, e.g. Optional.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum TypeHandle {
     Direct,
@@ -568,6 +561,9 @@ pub enum TypeHandle {
     VarArgs,
 }
 
+/// Used by sl_sh_fn macro to embed information at runtime about the parameters of
+/// the rust native function, specifically whether it is going to pass the value (a move),
+/// a reference, or mutable reference.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub enum PassingStyle {
     Value,
@@ -575,1369 +571,10 @@ pub enum PassingStyle {
     MutReference,
 }
 
+/// Struct used by sl_sh_fn macro to embed information in an array at runtime about each of
+/// the parameters of the rust native function.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 pub struct Param {
     pub handle: TypeHandle,
     pub passing_style: PassingStyle,
-}
-
-pub fn has_optional_params(params: &[Param]) -> bool {
-    for p in params {
-        match p.handle {
-            TypeHandle::Direct => {}
-            TypeHandle::Optional => {
-                return true;
-            }
-            TypeHandle::VarArgs => {
-                return true;
-            }
-        }
-    }
-    false
-}
-
-pub fn num_required_args(params: &[Param]) -> usize {
-    params.iter().fold(0, |accum, nxt| {
-        if nxt.handle == TypeHandle::Direct {
-            accum + 1
-        } else {
-            accum
-        }
-    })
-}
-
-/// arrays may be of different sizes, but there is at least 1 param that is optional or vec.
-fn get_args_optional_aware(
-    fn_name: &str,
-    params: Vec<Param>,
-    args: Vec<Expression>,
-) -> LispResult<Vec<ArgType>> {
-    let mut parsed_args = vec![];
-    // if they are equal it should be easy to parse regardless of the types, it's one for one.
-    if args.len() == params.len() {
-        for (exp, arg) in args.into_iter().zip(params.into_iter()) {
-            match arg.handle {
-                TypeHandle::Direct => parsed_args.push(ArgType::Exp(exp)),
-                TypeHandle::Optional => parsed_args.push(ArgType::Opt(Some(exp))),
-                TypeHandle::VarArgs => parsed_args.push(ArgType::VarArgs(vec![exp])),
-            }
-        }
-    } else {
-        // if there are more args than params for the receiving fn,
-        // the last param must be a vector.
-        if args.len() > params.len() {
-            if let Some(param) = params.iter().rev().next() {
-                match param.handle {
-                    TypeHandle::Direct | TypeHandle::Optional => {
-                        return Err(LispError::new(format!(
-                            "{} given too many arguments, expected {}, got {}.",
-                            fn_name,
-                            params.len(),
-                            args.len()
-                        )));
-                    }
-                    TypeHandle::VarArgs => {}
-                }
-            }
-        }
-        let mut args_iter = args.into_iter();
-        for param in params {
-            if let Some(exp) = args_iter.next() {
-                match param.handle {
-                    TypeHandle::Direct => parsed_args.push(ArgType::Exp(exp)),
-                    TypeHandle::Optional => parsed_args.push(ArgType::Opt(Some(exp))),
-                    // There can only be one TypeHandle::VarArgs and it's always the last one, this is
-                    // enforced at compile time.
-                    TypeHandle::VarArgs => {
-                        let mut exps = vec![exp];
-                        for exp in args_iter.by_ref() {
-                            exps.push(exp);
-                        }
-                        parsed_args.push(ArgType::VarArgs(exps));
-                    }
-                }
-            } else {
-                match param.handle {
-                    TypeHandle::Direct => {
-                        // it can't be the case that we don't have enough required arguments,
-                        // we already checked for that.
-                    }
-                    TypeHandle::Optional => parsed_args.push(ArgType::Opt(None)),
-                    TypeHandle::VarArgs => parsed_args.push(ArgType::VarArgs(vec![])),
-                }
-            }
-        }
-    }
-    Ok(parsed_args)
-}
-
-///// arrays may be of different sizes, but there is at least 1 param that is optional or vec.
-//fn verify_args_optional_aware(
-//    fn_name: &str,
-//    params: &[Arg],
-//    args: Vec<Expression>,
-//) -> LispResult<()> {
-//    // if there are more args than params for the receiving fn,
-//    // the last param must be a vector.
-//    Ok(())
-//}
-//
-///// arrays may be of different sizes, but there is at least 1 param that is optional or vec.
-//pub fn get_args_optional_aware_fast(
-//    fn_name: &str,
-//    params: Vec<Arg>,
-//    args: impl Iterator<Item = LispResult<Expression>> + 'static,
-//) -> LispResult<impl Iterator<Item = LispResult<ArgType>> + '_> {
-//    // if they are equal it should be easy to parse regardless of the types, it's one for one.
-//    let mut args = args.enumerate();
-//    let required_args = num_required_args(params.as_slice());
-//    let has_optional = has_optional_params(params.as_slice());
-//    let has_optional_str = if has_optional { "at least " } else { "" };
-//    let params_len = params.len();
-//    Ok(params.into_iter().enumerate().map(move |(i, param)| {
-//        if let Some((j, nxt)) = args.next() {
-//            if j > i && param.handle != TypeHandle::VarArgs {
-//                return Err(LispError::new(format!(
-//                    "{} given too many arguments, expected {}, got {}.",
-//                    fn_name,
-//                    params_len,
-//                    0, //TODO fix me, need global count.
-//                )));
-//            } else if j >= i && param.handle == TypeHandle::VarArgs {
-//                let capacity = j - i;
-//                let mut var_args = Vec::with_capacity(capacity);
-//                let mut nxt = nxt?;
-//                while j >= i {
-//                    var_args.push(nxt.clone());
-//                    if let Some((j, new_nxt)) = args.next() {
-//                        nxt = new_nxt?;
-//                    }
-//                }
-//                Ok(ArgType::VarArgs(var_args))
-//            } else {
-//                match param.handle {
-//                    TypeHandle::Value => {
-//                        let nxt = nxt?;
-//                        Ok(ArgType::Exp(nxt))
-//                    }
-//                    TypeHandle::Optional => {
-//                        let nxt = nxt?;
-//                        Ok(ArgType::Opt(Some(nxt)))
-//                    }
-//                    TypeHandle::VarArgs => {
-//                        unreachable!(
-//                            "VarArgs can only be last argument, this is checked at compile time."
-//                        )
-//                    }
-//                }
-//            }
-//        } else {
-//            return Err(LispError::new(format!(
-//                "{} given ??? arguments, expected {}, got {}.",
-//                fn_name,
-//                params_len,
-//                0, //TODO fix me, need global count.
-//            )));
-//        }
-//    }))
-//}
-
-/// convert function arguments to ArgTypes so call_expand_args gets
-/// what it expects
-pub fn get_arg_types(
-    fn_name: &str,
-    params: Vec<Param>,
-    args: Vec<Expression>,
-) -> LispResult<Vec<ArgType>> {
-    let parsed_args = vec![];
-    let required_args = num_required_args(params.as_slice());
-    let has_optional = has_optional_params(params.as_slice());
-    let has_optional_str = if has_optional { "at least " } else { "" };
-    match (
-        args.len() < required_args,
-        has_optional,
-        args.len() == required_args,
-        args.len() > required_args,
-    ) {
-        (true, _, _, _) => {
-            return Err(LispError::new(format!(
-                "{} not given enough arguments, expected {}{}, got {}.",
-                fn_name,
-                has_optional_str,
-                required_args,
-                args.len()
-            )));
-        }
-        (_, true, _, _) | (_, _, true, _) => {
-            return get_args_optional_aware(fn_name, params, args);
-        }
-        (_, _, _, true) => {
-            // if the length of the passed in args is
-            return Err(LispError::new(format!(
-                "{} given too many arguments, expected {}, got {}.",
-                fn_name,
-                required_args,
-                args.len()
-            )));
-        }
-        (false, false, false, false) => {}
-    }
-    Ok(parsed_args)
-}
-
-/// convert function arguments to ArgTypes so call_expand_args gets
-/// what it expects
-pub fn validate_args(fn_name: &str, params: &[Param], args: &[Expression]) -> LispResult<()> {
-    let required_args = num_required_args(params);
-    let has_optional = has_optional_params(params);
-    let has_optional_str = if has_optional { "at least " } else { "" };
-    // this match is about determining how the number of provided arguments
-    // compares to the number of required arguments (ArgType::Value)  along
-    // with knowledge of whether or not optional arguments exist.
-    match (
-        args.len() < required_args,
-        has_optional,
-        args.len() == required_args,
-        args.len() > required_args,
-    ) {
-        (true, _, _, _) => Err(LispError::new(format!(
-            "{} not given enough arguments, expected {}{}, got {}.",
-            fn_name,
-            has_optional_str,
-            required_args,
-            args.len()
-        ))),
-        (_, _, true, _) => Ok(()), // if the provided arguments == the required arguments, parsing is easy, validated.
-        (_, true, _, true) => {
-            // given there are optional parameters and the number of arguments exceeds the number
-            // of required ones, compare the arity of the target rust function with the number
-            // of provided arguments.
-            match args.len().cmp(&params.len()) {
-                Ordering::Less => Err(LispError::new(format!(
-                    "{} given too many arguments, expected {}, got {}.",
-                    fn_name,
-                    params.len(),
-                    args.len()
-                ))),
-                Ordering::Equal => Ok(()),
-                Ordering::Greater => {
-                    // this means the user passed too many arguments, unless,
-                    // the last param is a Vec type, which means args.len() > params.len()
-                    // is permitted as args can be arbitrarily large in this scenario.
-                    if let Some(param) = params.iter().rev().next() {
-                        match param.handle {
-                            TypeHandle::Direct | TypeHandle::Optional => {
-                                Err(LispError::new(format!(
-                                    "{} given too many arguments, expected {}, got {}.",
-                                    fn_name,
-                                    params.len(),
-                                    args.len()
-                                )))
-                            }
-                            TypeHandle::VarArgs => Ok(()),
-                        }
-                    } else {
-                        Err(LispError::new(format!(
-                            "{} given too many arguments, expected {}, got {}.",
-                            fn_name,
-                            params.len(),
-                            args.len()
-                        )))
-                    }
-                }
-            }
-        }
-        (_, false, _, true) => {
-            // if the length of the passed in args is greater than the number of required arguments
-            // with no optional parameters then the last argument must be a VarArgs as long as the
-            // parameter rules are being properly checked at compile time.
-            if let Some(param) = params.iter().rev().next() {
-                match param.handle {
-                    TypeHandle::Direct | TypeHandle::Optional => Err(LispError::new(format!(
-                        "{} given too many arguments, expected {}, got {}.",
-                        fn_name,
-                        params.len(),
-                        args.len()
-                    ))),
-                    TypeHandle::VarArgs => Ok(()),
-                }
-            } else {
-                Err(LispError::new(format!(
-                    "{} given too many arguments, expected {}, got {}.",
-                    fn_name,
-                    params.len(),
-                    args.len()
-                )))
-            }
-        }
-        (false, _, false, false) => unreachable!(
-            "The args slice can't be not less than, greater than, or equal to required_args"
-        ),
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::LispResult;
-    use std::collections::HashMap;
-    use std::convert::TryFrom;
-    use std::convert::TryInto;
-
-    fn loop_over_to_inline<const N: usize>(
-        fn_name: &str,
-        params: &[Param; N],
-        args: &[Expression],
-    ) -> LispResult<()> {
-        let required_args = num_required_args(params);
-        for idx in 0..N {
-            to_inline(fn_name, idx, required_args, params, args)?;
-        }
-        if N > 0 {
-            too_many_args_detection(fn_name, params, args)?;
-        }
-
-        Ok(())
-    }
-
-    fn too_many_args_detection<const N: usize>(
-        fn_name: &str,
-        params: &[Param; N],
-        args: &[Expression],
-    ) -> LispResult<()> {
-        let last_param = params[N - 1];
-        match args.get(N) {
-            Some(_) if last_param.handle != TypeHandle::VarArgs => {
-                return Err(LispError::new(format!(
-                    "{} given too many arguments, expected {}, got {}.",
-                    fn_name,
-                    params.len(),
-                    args.len()
-                )));
-            }
-            _ => {
-                //macro
-                println!("macro")
-            }
-        }
-        Ok(())
-    }
-
-    fn to_inline<const N: usize>(
-        fn_name: &str,
-        idx: usize,
-        required_args: usize,
-        params: &[Param; N],
-        args: &[Expression],
-    ) -> LispResult<()> {
-        let param = params[idx];
-        match args.get(idx) {
-            None if param.handle == TypeHandle::Direct => {
-                return Err(LispError::new(format!(
-                    "{} not given enough arguments, expected at least {} arguments, got {}.",
-                    fn_name,
-                    required_args,
-                    args.len()
-                )));
-            }
-            _arg => {
-                // insert
-                println!("macro");
-            }
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn test_params_values_only() {
-        let two_moved_values = vec![
-            Param {
-                handle: TypeHandle::Direct,
-                passing_style: PassingStyle::Value,
-            },
-            Param {
-                handle: TypeHandle::Direct,
-                passing_style: PassingStyle::Value,
-            },
-        ];
-
-        // if there are not enough arguments we throw an error.
-        let args = vec![Expression::make_true()];
-        let args = loop_over_to_inline::<2>(
-            "foo",
-            two_moved_values.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        );
-        assert!(args
-            .unwrap_err()
-            .reason
-            .contains("not given enough arguments"));
-
-        // if there are too many arguments we throw an error.
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        let args = loop_over_to_inline::<2>(
-            "foo",
-            two_moved_values.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        );
-        assert!(args.unwrap_err().reason.contains("given too many"));
-
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        loop_over_to_inline::<2>(
-            "foo",
-            two_moved_values.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-    }
-
-    #[test]
-    fn test_params_optionals() {
-        let one_val_one_opt = vec![
-            Param {
-                handle: TypeHandle::Direct,
-                passing_style: PassingStyle::Value,
-            },
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::Value,
-            },
-        ];
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        loop_over_to_inline::<2>(
-            "foo",
-            one_val_one_opt.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-
-        let args = vec![Expression::make_true()];
-        loop_over_to_inline::<2>(
-            "foo",
-            one_val_one_opt.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-
-        let args = vec![];
-        let args = loop_over_to_inline::<2>(
-            "foo",
-            one_val_one_opt.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        );
-        assert!(args
-            .unwrap_err()
-            .reason
-            .contains("not given enough arguments"));
-
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        let args = loop_over_to_inline::<2>(
-            "foo",
-            one_val_one_opt.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        );
-        assert!(args.unwrap_err().reason.contains("given too many"));
-
-        let val_and_opt = vec![
-            Param {
-                handle: TypeHandle::Direct,
-                passing_style: PassingStyle::Value,
-            },
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::Value,
-            },
-        ];
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        loop_over_to_inline::<2>(
-            "foo",
-            val_and_opt.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-
-        let args = vec![Expression::make_true()];
-        loop_over_to_inline::<2>(
-            "foo",
-            val_and_opt.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-
-        let args = vec![];
-        let args = loop_over_to_inline::<2>(
-            "foo",
-            val_and_opt.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        );
-        assert!(args
-            .unwrap_err()
-            .reason
-            .contains("not given enough arguments"));
-
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        let args = loop_over_to_inline::<2>(
-            "foo",
-            val_and_opt.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        );
-        assert!(args.unwrap_err().reason.contains("given too many"));
-    }
-
-    #[test]
-    fn test_params_vec() {
-        let one_vec = vec![Param {
-            handle: TypeHandle::VarArgs,
-            passing_style: PassingStyle::MutReference,
-        }];
-
-        let args = vec![];
-        loop_over_to_inline::<1>(
-            "foo",
-            one_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-
-        let args = vec![Expression::make_true()];
-        loop_over_to_inline::<1>(
-            "foo",
-            one_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        loop_over_to_inline::<1>(
-            "foo",
-            one_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        loop_over_to_inline::<1>(
-            "foo",
-            one_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-    }
-
-    #[test]
-    fn test_params_vec_with_options() {
-        let val_opt_and_vec = vec![
-            Param {
-                handle: TypeHandle::Direct,
-                passing_style: PassingStyle::Reference,
-            },
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::MutReference,
-            },
-            Param {
-                handle: TypeHandle::VarArgs,
-                passing_style: PassingStyle::Value,
-            },
-        ];
-
-        let args = vec![];
-        let args = loop_over_to_inline::<3>(
-            "foo",
-            val_opt_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        );
-        assert!(args
-            .unwrap_err()
-            .reason
-            .contains("not given enough arguments"));
-        let args = vec![Expression::make_true()];
-        loop_over_to_inline::<3>(
-            "foo",
-            val_opt_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        loop_over_to_inline::<3>(
-            "foo",
-            val_opt_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        loop_over_to_inline::<3>(
-            "foo",
-            val_opt_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        loop_over_to_inline::<3>(
-            "foo",
-            val_opt_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-
-        let opts_and_vec = vec![
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::Reference,
-            },
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::MutReference,
-            },
-            Param {
-                handle: TypeHandle::VarArgs,
-                passing_style: PassingStyle::Value,
-            },
-        ];
-
-        let args = vec![];
-        loop_over_to_inline::<3>(
-            "foo",
-            opts_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-        let args = vec![Expression::make_true()];
-        loop_over_to_inline::<3>(
-            "foo",
-            opts_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        loop_over_to_inline::<3>(
-            "foo",
-            opts_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        loop_over_to_inline::<3>(
-            "foo",
-            opts_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        loop_over_to_inline::<3>(
-            "foo",
-            opts_and_vec.as_slice().try_into().unwrap(),
-            args.as_slice(),
-        )
-        .expect("Parsing should succeed.");
-    }
-
-    #[test]
-    fn test_from_numerical() {
-        let result = Expression::from(42i64);
-        assert_eq!(result, Expression::alloc_data(ExpEnum::Int(42i64)));
-
-        let result: Expression = 42i64.into();
-        assert_eq!(result, Expression::alloc_data(ExpEnum::Int(42i64)));
-
-        let result = i64::try_from(Expression::alloc_data(ExpEnum::Int(7))).unwrap();
-        assert_eq!(7, result);
-
-        let result: f64 = Expression::alloc_data(ExpEnum::Int(7)).try_into().unwrap();
-        assert_eq!(7f64, result);
-
-        let result: f64 = Expression::alloc_data(ExpEnum::Float(7f64))
-            .try_into()
-            .unwrap();
-        assert_eq!(7f64, result);
-
-        let result: f64 = Expression::alloc_data(ExpEnum::Float(7f64))
-            .try_into()
-            .unwrap();
-        assert_eq!(7f64, result);
-
-        let result = f64::try_from(Expression::alloc_data(ExpEnum::Int(7i64))).unwrap();
-        assert_eq!(7f64, result);
-        let mut vs = vec![];
-        vs.push(vec![1, 2, 3]);
-        vs.push(vec![2, 2, 3]);
-        vs.push(vec![1, 2, 1]);
-        for vs in vs.iter().map(|v| v.iter()) {
-            for v in vs {
-                println!("V: {}", v);
-            }
-        }
-    }
-
-    impl PartialEq<Self> for Expression {
-        fn eq(&self, other: &Self) -> bool {
-            match (&self.get().data, &other.get().data) {
-                (ExpEnum::True, ExpEnum::True) => true,
-                (ExpEnum::False, ExpEnum::False) => true,
-                (ExpEnum::Nil, ExpEnum::Nil) => true,
-                (ExpEnum::Float(lf), ExpEnum::Float(rt)) => lf == rt,
-                (ExpEnum::Int(lf), ExpEnum::Int(rt)) => lf == rt,
-                (ExpEnum::Char(lf), ExpEnum::Char(rt)) => lf == rt,
-                (ExpEnum::CodePoint(lf), ExpEnum::CodePoint(rt)) => lf == rt,
-                (_, _) => false,
-            }
-            //(ExpEnum::String(lf_s, lf_i), ExpEnum::String(rt_s, rt_i) => {}
-            //(ExpEnum::Symbol(_, _), ExpEnum::Symbol(_, _)) => {}
-            //(ExpEnum::Regex(_), ExpEnum::Regex(_)) => {}
-            //ExpEnum::Symbol(_, _) => {}
-            //ExpEnum::Lambda(_) => {}
-            //ExpEnum::Macro(_) => {}
-            //ExpEnum::Function(_) => {}
-            //ExpEnum::LazyFn(_, _) => {}
-            //ExpEnum::Vector(_) => {}
-            //ExpEnum::Values(_) => {}
-            //ExpEnum::Pair(_, _) => {}
-            //ExpEnum::HashMap(_) => {}
-            //ExpEnum::Process(_) => {}
-            //ExpEnum::File(_) => {}
-            //ExpEnum::Wrapper(_) => {}
-            //ExpEnum::DeclareDef => {}
-            //ExpEnum::DeclareVar => {}
-            //ExpEnum::DeclareFn => {}
-            //ExpEnum::DeclareMacro => {}
-            //ExpEnum::Quote => {}
-            //ExpEnum::BackQuote => {}
-            //ExpEnum::Undefined => {}
-        }
-    }
-
-    static K0: &'static str = "key0";
-    static K1: &'static str = "key1";
-
-    #[test]
-    fn with_some_hashmaps_test() {
-        let mut hash_map = HashMap::new();
-        hash_map.insert(K0, Expression::alloc_data(ExpEnum::Int(7)));
-        hash_map.insert(K1, Expression::alloc_data(ExpEnum::Int(11)));
-        let myint0 = ArgType::Exp(Expression::alloc_data(ExpEnum::HashMap(hash_map)));
-        arg_translate_clear_hash_map(myint0).unwrap();
-    }
-
-    fn arg_translate_clear_hash_map(arg_0: ArgType) -> crate::LispResult<()> {
-        try_exp_enum!(
-            arg_0,
-            ArgType::Exp(exp),
-            arg_unwrap_clear_hash_map(exp)?,
-            "err"
-        )
-    }
-
-    fn arg_unwrap_clear_hash_map(exp_0: Expression) -> crate::LispResult<()> {
-        try_exp_enum!(
-            exp_0.get_mut().data,
-            ExpEnum::HashMap(ref mut hash_map),
-            clear_hash_map(hash_map)?,
-            "Not an int_0!"
-        )
-    }
-
-    fn clear_hash_map(hash_map: &mut HashMap<&str, Expression>) -> LispResult<()> {
-        hash_map.clear();
-        Ok(())
-    }
-
-    #[test]
-    fn with_some_optional_hashmaps_test() {
-        let mut hash_map = HashMap::new();
-        hash_map.insert(K0, Expression::alloc_data(ExpEnum::Int(7)));
-        hash_map.insert(K1, Expression::alloc_data(ExpEnum::Int(11)));
-        let myint0 = ArgType::Opt(Some(Expression::alloc_data(ExpEnum::HashMap(hash_map))));
-        arg_translate_clear_hash_map_optional(myint0).unwrap();
-    }
-
-    fn arg_translate_clear_hash_map_optional(arg_0: ArgType) -> crate::LispResult<()> {
-        try_exp_enum!(
-            arg_0,
-            ArgType::Opt(exp),
-            arg_unwrap_clear_hash_map_optional(exp)?,
-            "Was not parseable as an option"
-        )
-    }
-
-    fn arg_unwrap_clear_hash_map_optional(exp_0: Option<Expression>) -> crate::LispResult<()> {
-        match exp_0 {
-            None => Ok(clear_hash_map_optional(None)?),
-            Some(exp_0) => {
-                try_exp_enum!(
-                    exp_0.get_mut().data,
-                    ExpEnum::HashMap(ref mut hash_map),
-                    clear_hash_map_optional(Some(hash_map))?,
-                    "Not an int_0!"
-                )
-            }
-        }
-    }
-
-    fn clear_hash_map_optional(hash_map: Option<&mut HashMap<&str, Expression>>) -> LispResult<()> {
-        if let Some(hash_map) = hash_map {
-            hash_map.clear();
-        }
-        Ok(())
-    }
-
-    #[test]
-    fn int2float_negative_test() {
-        let myint0 = ArgType::Exp(Expression::alloc_data(ExpEnum::Float(7.0)));
-        let myint1 = ArgType::Exp(Expression::alloc_data(ExpEnum::Int(11)));
-        let result = arg_translate_int_2_float(myint0, myint1);
-        match result {
-            Ok(_) => {
-                panic!("result should not be ok!");
-            }
-            Err(e) => {
-                assert_eq!(e.reason, "Not an int_0!");
-            }
-        }
-        let myint0 = ArgType::Exp(Expression::alloc_data(ExpEnum::Int(7)));
-        let myint1 = ArgType::Exp(Expression::alloc_data(ExpEnum::Float(11.0)));
-        let result = arg_translate_int_2_float(myint0, myint1);
-        match result {
-            Ok(_) => {
-                panic!("result should not be ok!");
-            }
-            Err(e) => {
-                assert_eq!(e.reason, "Not an int_1!");
-            }
-        }
-    }
-
-    #[test]
-    fn int2float_test() {
-        let myint0 = ArgType::Exp(Expression::alloc_data(ExpEnum::Int(7)));
-        let myint1 = ArgType::Exp(Expression::alloc_data(ExpEnum::Int(11)));
-        let result = arg_translate_int_2_float(myint0, myint1).unwrap();
-        let res_d = &result.get().data;
-        match res_d {
-            ExpEnum::Float(f) => {
-                assert_eq!(18.0, *f);
-            }
-            _ => {
-                panic!("Not a float!")
-            }
-        }
-    }
-
-    fn arg_translate_int_2_float(arg_0: ArgType, arg_1: ArgType) -> crate::LispResult<Expression> {
-        try_exp_enum!(
-            arg_0,
-            ArgType::Exp(exp_0),
-            {
-                try_exp_enum!(
-                    arg_1,
-                    ArgType::Exp(exp_1),
-                    { arg_unwrap_int_2_float(exp_0, exp_1).map(Into::into) }?,
-                    "err"
-                )
-            }?,
-            "err exp_0"
-        )
-    }
-
-    fn arg_unwrap_int_2_float(exp_0: Expression, exp_1: Expression) -> crate::LispResult<f64> {
-        Ok({
-            let float = ret_err_exp_enum!(
-                exp_0.get().data,
-                ExpEnum::Int(int_0),
-                {
-                    let float = ret_err_exp_enum!(
-                        exp_1.get().data,
-                        ExpEnum::Int(int_1),
-                        { int_2_float(int_0, int_1)? },
-                        "Not an int_1!"
-                    );
-                    float
-                },
-                "Not an int_0!"
-            );
-            float
-        })
-    }
-
-    fn int_2_float(my_int: i64, my_o_int: i64) -> LispResult<f64> {
-        Ok((my_int + my_o_int) as f64)
-    }
-
-    #[test]
-    fn optional_int2float_test() {
-        let myint0 = ArgType::Exp(Expression::alloc_data(ExpEnum::Int(7)));
-        let myint1 = ArgType::Opt(Some(Expression::alloc_data(ExpEnum::Int(11))));
-        let result = arg_translate_optional_int_2_float(myint0, myint1).unwrap();
-        let res_d = &result.get().data;
-        match res_d {
-            ExpEnum::Float(f) => {
-                assert_eq!(18.0, *f);
-            }
-            _ => {
-                panic!("Not a float!")
-            }
-        }
-
-        let myint0 = ArgType::Exp(Expression::alloc_data(ExpEnum::Int(7)));
-        let myint1 = ArgType::Opt(None);
-        let result = arg_translate_optional_int_2_float(myint0, myint1).unwrap();
-        let res_d = &result.get().data;
-        match res_d {
-            ExpEnum::Float(f) => {
-                assert_eq!(7.0, *f);
-            }
-            _ => {
-                panic!("Not a float!")
-            }
-        }
-    }
-
-    fn arg_translate_optional_int_2_float(
-        arg_0: ArgType,
-        arg_1: ArgType,
-    ) -> crate::LispResult<Expression> {
-        let exp_0 = ret_err_exp_enum!(arg_0, ArgType::Exp(exp), exp, "err");
-        let exp_1 = ret_err_exp_enum!(arg_1, ArgType::Opt(exp), exp, "err");
-        arg_unwrap_optional_int_2_float(exp_0, exp_1).map(Into::into)
-    }
-
-    fn arg_unwrap_optional_int_2_float(
-        exp_0: Expression,
-        exp_1: Option<Expression>,
-    ) -> crate::LispResult<f64> {
-        Ok({
-            let float = ret_err_exp_enum!(
-                exp_0.get().data,
-                ExpEnum::Int(int_0),
-                {
-                    match exp_1 {
-                        None => {
-                            let int_1 = None;
-                            optional_int_2_float(int_0, int_1)
-                        }
-                        Some(exp_1) => {
-                            let float = ret_err_exp_enum!(
-                                exp_1.get().data,
-                                ExpEnum::Int(int_1),
-                                { optional_int_2_float(int_0, Some(int_1)) },
-                                "Not an int_1!"
-                            );
-                            float
-                        }
-                    }
-                },
-                "Not an int_0!"
-            );
-            float
-        })
-    }
-
-    fn optional_int_2_float(my_int: i64, my_o_int: Option<i64>) -> f64 {
-        if let Some(my_o_int) = my_o_int {
-            (my_int + my_o_int) as f64
-        } else {
-            my_int as f64
-        }
-    }
-
-    #[test]
-    fn ints_2float_test() {
-        let myint0 = ArgType::Exp(Expression::alloc_data(ExpEnum::Int(7)));
-        let myint1 = ArgType::VarArgs(vec![
-            Expression::alloc_data(ExpEnum::Int(11)),
-            Expression::alloc_data(ExpEnum::Int(2)),
-        ]);
-        let result = arg_translate_ints_2_float(myint0, myint1).unwrap();
-        let res_d = &result.get().data;
-        match res_d {
-            ExpEnum::Float(f) => {
-                assert_eq!(20.0, *f);
-            }
-            _ => {
-                panic!("Not a float!")
-            }
-        }
-    }
-
-    fn arg_translate_ints_2_float(arg_0: ArgType, arg_1: ArgType) -> crate::LispResult<Expression> {
-        let exp_0 = ret_err_exp_enum!(arg_0, ArgType::Exp(exp), exp, "err");
-        let exp_1 = ret_err_exp_enum!(arg_1, ArgType::VarArgs(exp), exp, "err");
-        arg_unwrap_ints_2_float(exp_0, exp_1).map(Into::into)
-    }
-
-    fn arg_unwrap_ints_2_float(
-        exp_0: Expression,
-        exp_1: Vec<Expression>,
-    ) -> crate::LispResult<f64> {
-        Ok({
-            let float = ret_err_exp_enum!(
-                exp_0.get().data,
-                ExpEnum::Int(int_0),
-                {
-                    let iter = exp_1
-                        .iter()
-                        .map(|exp_1| {
-                            let int = ret_err_exp_enum!(
-                                exp_1.get().data,
-                                ExpEnum::Int(int_1),
-                                { int_1 },
-                                "Not an int_1!"
-                            );
-                            Ok(int)
-                        })
-                        .collect::<crate::LispResult<Vec<i64>>>()?;
-                    ints_2_float(int_0, iter)
-                },
-                "Not an int_0!"
-            );
-            float
-        })
-    }
-
-    fn ints_2_float(my_int: i64, my_o_ints: Vec<i64>) -> f64 {
-        my_o_ints
-            .iter()
-            .fold(my_int as f64, |sum, val| sum + *val as f64)
-    }
-
-    #[test]
-    fn test_builtin_int_2_float() {
-        let arg_0 = Expression::alloc_data(ExpEnum::Int(8));
-        let arg_1 = Expression::alloc_data(ExpEnum::Int(8));
-        let exp = sample_builtin_int_2_float(ArgType::Exp(arg_0), ArgType::Exp(arg_1)).unwrap();
-        let exp_d = exp.get();
-        match exp_d.data {
-            ExpEnum::Float(f) => {
-                assert_eq!(16.0, f);
-            }
-            _ => {
-                panic!("Should be float.");
-            }
-        }
-    }
-
-    fn sample_builtin_int_2_float(arg_0: ArgType, arg_1: ArgType) -> Result<Expression, LispError> {
-        match arg_1 {
-            ArgType::Exp (arg_1) => match arg_1.get().data {
-                ExpEnum::Int(arg_1) => {
-                    match arg_0 {
-                        ArgType::Exp(arg_0) => match arg_0.get().data {
-                            ExpEnum::Int (arg_0) => {
-                                int_2_float (arg_0 , arg_1).map(Into :: into)
-                            },
-                            _ => return Err (LispError :: new ("sl_sh_fn macro is broken, apparently ArgType::Exp can't be parsed as ArgType::Exp")) ,
-                        },
-                        _ => return Err (LispError :: new ("sl_sh_fn macro is broken, apparently ArgType::Exp can't be parsed as ArgType::Exp")) ,
-                    }
-                },
-                _ => return Err (LispError :: new ("sl_sh_fn macro is broken, apparently ArgType::Exp can't be parsed as ArgType::Exp")),
-            },
-            _ => return Err (LispError :: new ("sl_sh_fn macro is broken, apparently ArgType::Exp can't be parsed as ArgType::Exp")) ,
-        }
-    }
-
-    #[test]
-    fn test_vecs() {
-        let one_vec = vec![Param {
-            handle: TypeHandle::VarArgs,
-            passing_style: PassingStyle::MutReference,
-        }];
-        let args = vec![];
-        let args = get_arg_types("foo", one_vec.clone(), args).unwrap();
-        assert_eq!(1, args.len());
-        try_exp_enum!(
-            args.get(0).unwrap(),
-            ArgType::VarArgs(val),
-            assert!(val.is_empty()),
-            "err"
-        )
-        .unwrap();
-
-        let args = vec![Expression::make_true()];
-        let args = get_arg_types("foo", one_vec.clone(), args).unwrap();
-        assert_eq!(1, args.len());
-        try_exp_enum!(
-            args.get(0).unwrap(),
-            ArgType::VarArgs(val),
-            assert_eq!(1, val.len()),
-            "err"
-        )
-        .unwrap();
-
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        let args = get_arg_types("foo", one_vec.clone(), args).unwrap();
-        assert_eq!(1, args.len());
-        try_exp_enum!(
-            args.get(0).unwrap(),
-            ArgType::VarArgs(val),
-            assert_eq!(3, val.len()),
-            "err"
-        )
-        .unwrap();
-
-        let val_vec = vec![
-            Param {
-                handle: TypeHandle::Direct,
-                passing_style: PassingStyle::Reference,
-            },
-            Param {
-                handle: TypeHandle::VarArgs,
-                passing_style: PassingStyle::MutReference,
-            },
-        ];
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        let args = get_arg_types("foo", val_vec.clone(), args).unwrap();
-        assert_eq!(2, args.len());
-        try_exp_enum!(
-            args.get(1).unwrap(),
-            ArgType::VarArgs(val),
-            assert_eq!(3, val.len()),
-            "err"
-        )
-        .unwrap();
-    }
-
-    #[test]
-    fn test_vec_with_optionals() {
-        let val_opt_and_vec = vec![
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::Reference,
-            },
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::MutReference,
-            },
-            Param {
-                handle: TypeHandle::VarArgs,
-                passing_style: PassingStyle::Value,
-            },
-        ];
-
-        // optional arguments manifest as Some
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        let args = get_arg_types("foo", val_opt_and_vec.clone(), args).unwrap();
-        assert_eq!(3, args.len());
-        try_exp_enum!(
-            args.get(0).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_some()),
-            "err"
-        )
-        .unwrap();
-        try_exp_enum!(
-            args.get(1).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_some()),
-            "err"
-        )
-        .unwrap();
-        try_exp_enum!(
-            args.get(2).unwrap(),
-            ArgType::VarArgs(val),
-            assert!(val.is_empty()),
-            "err"
-        )
-        .unwrap();
-
-        // optional arguments manifest as Some
-        let args = vec![];
-        let args = get_arg_types("foo", val_opt_and_vec, args).unwrap();
-        assert_eq!(3, args.len());
-        try_exp_enum!(
-            args.get(0).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_none()),
-            "err"
-        )
-        .unwrap();
-        try_exp_enum!(
-            args.get(1).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_none()),
-            "err"
-        )
-        .unwrap();
-        try_exp_enum!(
-            args.get(2).unwrap(),
-            ArgType::VarArgs(val),
-            assert!(val.is_empty()),
-            "err"
-        )
-        .unwrap();
-    }
-
-    #[test]
-    fn test_all_optional() {
-        let two_opts = vec![
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::Value,
-            },
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::Value,
-            },
-        ];
-
-        // optional arguments manifest as Some
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        let args = get_arg_types("foo", two_opts.clone(), args).unwrap();
-        assert_eq!(2, args.len());
-        try_exp_enum!(
-            args.get(0).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_some()),
-            "err"
-        )
-        .unwrap();
-        try_exp_enum!(
-            args.get(1).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_some()),
-            "err"
-        )
-        .unwrap();
-
-        // optional arguments manifest as Some
-        let args = vec![];
-        let args = get_arg_types("foo", two_opts, args).unwrap();
-        assert_eq!(2, args.len());
-        try_exp_enum!(
-            args.get(0).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_none()),
-            "err"
-        )
-        .unwrap();
-        try_exp_enum!(
-            args.get(1).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_none()),
-            "err"
-        )
-        .unwrap();
-    }
-
-    #[test]
-    fn test_optional() {
-        let one_val_one_opt = vec![
-            Param {
-                handle: TypeHandle::Direct,
-                passing_style: PassingStyle::Value,
-            },
-            Param {
-                handle: TypeHandle::Optional,
-                passing_style: PassingStyle::Value,
-            },
-        ];
-
-        // optional arguments manifest as Some
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        let args = get_arg_types("foo", one_val_one_opt.clone(), args).unwrap();
-        assert_eq!(2, args.len());
-        try_exp_enum!(args.get(0).unwrap(), ArgType::Exp(_), {}, "err").unwrap();
-        try_exp_enum!(
-            args.get(1).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_some()),
-            "err"
-        )
-        .unwrap();
-
-        // optional arguments are optional
-        let args = vec![Expression::make_true()];
-        let args = get_arg_types("foo", one_val_one_opt, args).unwrap();
-        assert_eq!(2, args.len());
-        try_exp_enum!(args.get(0).unwrap(), ArgType::Exp(_), {}, "err").unwrap();
-        try_exp_enum!(
-            args.get(1).unwrap(),
-            ArgType::Opt(val),
-            assert!(val.is_none()),
-            "err"
-        )
-        .unwrap();
-    }
-
-    #[test]
-    fn test_values() {
-        let two_moved_values = vec![
-            Param {
-                handle: TypeHandle::Direct,
-                passing_style: PassingStyle::Value,
-            },
-            Param {
-                handle: TypeHandle::Direct,
-                passing_style: PassingStyle::Value,
-            },
-        ];
-
-        // if there are not enough arguments we throw an error.
-        let args = vec![Expression::make_true()];
-        let args = get_arg_types("foo", two_moved_values.clone(), args);
-        assert!(args
-            .unwrap_err()
-            .reason
-            .contains("not given enough arguments"));
-
-        // if there are too many arguments we throw an error.
-        let args = vec![
-            Expression::make_true(),
-            Expression::make_true(),
-            Expression::make_true(),
-        ];
-        let args = get_arg_types("foo", two_moved_values.clone(), args);
-        assert!(args.unwrap_err().reason.contains("given too many"));
-
-        // the correct number of arguments succeeds.
-        let args = vec![Expression::make_true(), Expression::make_true()];
-        let args = get_arg_types("foo", two_moved_values, args).unwrap();
-        assert_eq!(2, args.len());
-        try_exp_enum!(args.get(0).unwrap(), ArgType::Exp(_), {}, "err").unwrap();
-        try_exp_enum!(args.get(1).unwrap(), ArgType::Exp(_), {}, "err").unwrap();
-    }
 }
